authservice/requirements.txt
bcrypt
certifi==2020.6.20
cffi
click
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-SQLAlchemy
itsdangerous==1.1.0
Jinja2
MarkupSafe==1.1.1
psycopg2
pycparser
PyJWT
python-dotenv
six
SQLAlchemy
Werkzeug==1.0.1
